requirements.txt
speedtest-cli==2.1.3
pandas>=2.0.0
schedule>=1.2.0
rich>=13.0.0
typer>=0.9.0